+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Get list of installable modules from pext.hackerchick.me instead of pext.github.io
+
 ## [0.3] - 2016-12-29
 ### API changes
 - The entry and command list will now be emptied each time just before selection_made is called