+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Basic Qt5 theming support using installed system themes
+
 ### Changed
 - Get list of installable modules from pext.hackerchick.me instead of pext.github.io